requirements.txt
matplotlib==3.5.0
pandas==1.3.4
streamlit==1.2.0
statsmodels==0.10.2
pillow==8.4.0